Implemented tests for controller, repository levels (#12)
and service for book and categorÐy

+++ src/main/java/book/store/service/book/BookServiceImpl.java
@@ -62,6 +62,10 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteBookById(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find book by id: " + id);
+        }
+
         bookRepository.deleteById(id);
     }
